add past checkouts

+++ accounting/views.py
@@ -4,7 +4,6 @@ from django.contrib import messages
 from django.contrib.auth import logout, authenticate, login
 from django.shortcuts import render, redirect
 
-from accounting.controller import cal_min_cash_flow
 from accounting.forms import *
 
 
@@ -299,25 +298,30 @@ def visit_group(request, group_id):
         for member in group.members.all():
             member.debt = 0
             member.save()
-
         expenses = Expense.objects.filter(group=group_id)
-        transactions = cal_min_cash_flow(expenses)
+        print(expenses)
         member_debts = {}
         my_debt = 0
-        for transaction in transactions:
-            receiver, payer, amount = transaction
-            if payer == request.user.username:
-                my_debt += amount
-            elif receiver == request.user.username:
-                member_debts[payer] = member_debts.get(payer, 0) + amount
+        for expense in expenses:
+            if expense.spender.username == request.user.username:
+                debts = Debt.objects.filter(expense=expense)
+                for debt in debts:
+                    print(debt.share, debt.person.first_name)
+                    if debt.share > 0:
+                        print(member_debts.keys(),debt.person.username)
+                        member_debts[debt.person.username] = member_debts.get(debt.person.username,0) + expense.cost*debt.share
+            else:
+                debts = Debt.objects.filter(expense=expense, person=request.user)
+                for debt in debts:
+                    my_debt += expense.cost*debt.share
 
         member = group.members.all().filter(username=request.user.username)[0]
         member.debt = my_debt
         member.save()
-        for username in member_debts.keys():
-            print(username, member_debts[username])
-            member = group.members.all().filter(username=username)[0]
-            member.debt = member_debts[username]
+        for member_debt in member_debts.keys():
+            print(member_debt,member_debts[member_debt])
+            member = group.members.all().filter(username=member_debt)[0]
+            member.debt = member_debts[member_debt]
             member.save()
 
         context = {
@@ -327,6 +331,7 @@ def visit_group(request, group_id):
             'members': group.members.all(),
             'expenses': expenses,
         }
+
         return render(request, 'group/visit_group.html', context=context)
     else:
         return redirect('/login/')
@@ -426,23 +431,29 @@ def confirm_expense(request, group_id):
     else:
         return redirect('/login/')
 
-
 def checkout_expense(request, group_id):
     if request.user.is_authenticated:
         if request.method == 'POST':
+            print(group_id)
             group = ExpenseGroup.objects.get(pk=group_id)
             for member in group.members.all():
                 member.debt = 0
                 member.save()
-
+            print(group.members.all())
             expenses = Expense.objects.filter(group=group_id)
-            transactions = cal_min_cash_flow(expenses)
-            for transaction in transactions:
-                receiver, payer, amount = transaction
-                if payer == request.user.username:
-                    receiver_user = User.objects.get_by_natural_key(receiver)
-                    receiver_user.debt += amount
-                    receiver_user.save()
+            member_debts = {}
+            for expense in expenses:
+                if expense.spender.username != request.user.username:
+                    debts = Debt.objects.filter(expense=expense, person=request.user)
+                    print("debtsdebts",debts)
+                    for debt in debts:
+                        member_debts[expense.spender.username] = member_debts.get(expense.spender.username,0) + expense.cost*debt.share
+
+            for member_debt in member_debts.keys():
+                print(member_debt,member_debts[member_debt])
+                member = group.members.all().filter(username=member_debt)[0]
+                member.debt = member_debts[member_debt]
+                member.save()
 
             context = {
                 'group_id': group_id,
@@ -453,32 +464,40 @@ def checkout_expense(request, group_id):
     else:
         return redirect('/login/')
 
-
 def confirm_checkout_expense(request, group_id, user_id):
     if request.user.is_authenticated:
         if request.method == 'POST':
             group = ExpenseGroup.objects.get(pk=group_id)
-
-            amount = float(request.POST['submit'])
-            receiver_user = User.objects.get(pk=user_id)
-            PastCheckouts.objects.create(cost=amount, payer=request.user, receiver=receiver_user)
-
-            receiver_user.debt = 0
-            receiver_user.save()
-
-            expenses = Expense.objects.filter(group=group_id)
-            while amount > 0:
-                expense = expenses.pop(0)
-                debt = Debt.objects.filter(expense=expense, person=request.user)[0]
-                amount -= min(amount, expense.cost * debt.share)
-                debt.delete()
+            member = group.members.all().filter(id=user_id)[0]
+            if member.debt >0:
+                for member in group.members.all():
+                    member.debt = 0
+                    member.save()
+                expenses = Expense.objects.filter(group=group_id)
+                for expense in expenses:
+                    print(expense.spender.id,user_id)
+                    if expense.spender.id == int(user_id):
+                        print("hereee")
+                        debts = Debt.objects.filter(expense=expense, person=request.user)
+                        for debt in debts:
+                            PastCheckouts.objects.create(cost=expense.cost*debt.share, payer=request.user, reciever= expense.spender)
+                            debt.delete()
+                    print("PastCheckouts",PastCheckouts.objects.get(pk=group_id))
+            else:
+                print("no debt")
 
             context = {
                 'group_id': group_id,
                 'members': group.members.all(),
             }
+
             return render(request, 'group/checkout_expense.html', context=context)
         else:
             return redirect('/visit_group/{}'.format(group_id))
     else:
         return redirect('/login/')
+
+def view_past_checkouts(request):
+    if request.user.is_authenticated:
+        return render(request, 'profile.html')
+    return redirect('/login/')
